requirements.txt
requests==2.32.3
tenacity==9.0.0
python-dotenv==1.0.1
pytest==8.3.3